# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tawsila Food Delivery API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  customer: "customer",
  driver: "driver",
  restaurant_owner: "restaurant_owner",
  admin: "admin",
} as const;

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
  role: RegisterBodyRole;
  /** @nullable */
  phone?: string | null;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  role: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  isActive: boolean;
  loyaltyPoints: number;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UpdateUserBody {
  name?: string;
  phone?: string;
  address?: string;
  avatarUrl?: string;
  isActive?: boolean;
}

export interface Restaurant {
  id: number;
  ownerId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  address: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
  category: string;
  isLocal: boolean;
  isOpen: boolean;
  /** @nullable */
  deliveryTime?: number | null;
  /** @nullable */
  deliveryFee?: number | null;
  /** @nullable */
  minimumOrder?: number | null;
  /** @nullable */
  rating?: number | null;
  reviewCount: number;
  isVerified: boolean;
  createdAt: string;
}

export interface CreateRestaurantBody {
  name: string;
  description?: string;
  address: string;
  phone?: string;
  imageUrl?: string;
  coverImageUrl?: string;
  category: string;
  isLocal?: boolean;
  deliveryTime?: number;
  deliveryFee?: number;
  minimumOrder?: number;
}

export interface UpdateRestaurantBody {
  name?: string;
  description?: string;
  address?: string;
  phone?: string;
  imageUrl?: string;
  coverImageUrl?: string;
  category?: string;
  isLocal?: boolean;
  isOpen?: boolean;
  deliveryTime?: number;
  deliveryFee?: number;
  minimumOrder?: number;
  isVerified?: boolean;
}

export interface RestaurantStats {
  totalOrders: number;
  completedOrders: number;
  totalRevenue: number;
  /** @nullable */
  averageRating?: number | null;
  totalReviews: number;
  pendingOrders: number;
}

export interface MenuItem {
  id: number;
  restaurantId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  imageUrl?: string | null;
  category: string;
  isAvailable: boolean;
  isPopular: boolean;
  createdAt: string;
}

export interface CreateMenuItemBody {
  name: string;
  description?: string;
  price: number;
  imageUrl?: string;
  category: string;
  isPopular?: boolean;
}

export interface UpdateMenuItemBody {
  name?: string;
  description?: string;
  price?: number;
  imageUrl?: string;
  category?: string;
  isAvailable?: boolean;
  isPopular?: boolean;
}

export interface OrderItem {
  id: number;
  orderId: number;
  menuItemId: number;
  menuItemName: string;
  quantity: number;
  unitPrice: number;
  totalPrice: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  accepted: "accepted",
  preparing: "preparing",
  ready: "ready",
  picked_up: "picked_up",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  userId: number;
  restaurantId: number;
  /** @nullable */
  driverId?: number | null;
  restaurantName: string;
  userName: string;
  status: OrderStatus;
  subtotal: number;
  deliveryFee: number;
  total: number;
  deliveryAddress: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  estimatedDeliveryTime?: number | null;
  items: OrderItem[];
  createdAt: string;
  updatedAt: string;
}

export type CreateOrderBodyItemsItem = {
  menuItemId: number;
  quantity: number;
};

export interface CreateOrderBody {
  restaurantId: number;
  deliveryAddress: string;
  notes?: string;
  items: CreateOrderBodyItemsItem[];
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  pending: "pending",
  accepted: "accepted",
  preparing: "preparing",
  ready: "ready",
  picked_up: "picked_up",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
  driverId?: number;
}

export interface Driver {
  id: number;
  userId: number;
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  vehicleType?: string | null;
  isAvailable: boolean;
  totalDeliveries: number;
  /** @nullable */
  rating?: number | null;
  createdAt: string;
}

export interface UpdateDriverBody {
  vehicleType?: string;
  isAvailable?: boolean;
}

export interface DriverEarnings {
  today: number;
  thisWeek: number;
  thisMonth: number;
  totalDeliveries: number;
  completedToday: number;
}

export interface Review {
  id: number;
  userId: number;
  restaurantId: number;
  /** @nullable */
  orderId?: number | null;
  userName: string;
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}

export interface CreateReviewBody {
  restaurantId: number;
  orderId?: number;
  rating: number;
  comment?: string;
}

export interface UserRewards {
  userId: number;
  loyaltyPoints: number;
  tier: string;
  nextTierPoints: number;
  totalOrdersCount: number;
  totalSpent: number;
}

export interface AdminStats {
  totalUsers: number;
  totalRestaurants: number;
  totalOrders: number;
  totalDrivers: number;
  activeOrders: number;
  revenue: number;
  ordersToday: number;
  newUsersToday: number;
}

export type ListRestaurantsParams = {
  search?: string;
  category?: string;
  isLocal?: boolean;
  isOpen?: boolean;
};

export type ListMenuItemsParams = {
  category?: string;
};

export type ListOrdersParams = {
  status?: string;
  userId?: number;
  restaurantId?: number;
  driverId?: number;
};

export type ListUsersParams = {
  role?: string;
  search?: string;
};

export type ListDriversParams = {
  isAvailable?: boolean;
};

export type ListReviewsParams = {
  restaurantId?: number;
  userId?: number;
};