# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tawsila Food Delivery API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /restaurants/{restaurantId}/menu:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List menu items for a restaurant
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"

    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Create a menu item
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMenuItemBody"
      responses:
        "201":
          description: Menu item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
  /menu/{id}:
    get:
      operationId: getMenuItem
      tags: [menu]
      summary: Get a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"

    patch:
      operationId: updateMenuItem
      tags: [menu]
      summary: Update a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMenuItemBody"
      responses:
        "200":
          description: Updated menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"

    delete:
      operationId: deleteMenuItem
      tags: [menu]
      summary: Delete a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Orders
...
  /drivers:
    get:
      operationId: listDrivers
      tags: [drivers]
      summary: List drivers
      parameters:
        - name: isAvailable
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Drivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Driver"
  /drivers/{id}:
    get:
      operationId: getDriver
      tags: [drivers]
      summary: Get a driver
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"

    patch:
      operationId: updateDriver
      tags: [drivers]
      summary: Update driver info
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDriverBody"
      responses:
        "200":
          description: Updated driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
  /drivers/{id}/earnings:
    get:
      operationId: getDriverEarnings
      tags: [drivers]
      summary: Get driver earnings summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Driver earnings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverEarnings"

  # Reviews
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [customer, driver, restaurant_owner, admin]
        phone:
          type: ["string", "null"]
      required:
        - name
        - email
        - password
        - role
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user

    # User
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        loyaltyPoints:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - isActive
        - loyaltyPoints
        - createdAt
...
    MenuItem:
      type: object
      properties:
        id:
          type: integer
        restaurantId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        category:
          type: string
        isAvailable:
          type: boolean
        isPopular:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - restaurantId
        - name
        - price
        - category
        - isAvailable
        - isPopular
        - createdAt
    CreateMenuItemBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        category:
          type: string
        isPopular:
          type: boolean
      required:
        - name
        - price
        - category
    UpdateMenuItemBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        category:
          type: string
        isAvailable:
          type: boolean
        isPopular:
          type: boolean

    # Order
...
    Driver:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        vehicleType:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        totalDeliveries:
          type: integer
        rating:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - isAvailable
        - totalDeliveries
        - createdAt
    UpdateDriverBody:
      type: object
      properties:
        vehicleType:
          type: string
        isAvailable:
          type: boolean
    DriverEarnings:
      type: object
      properties:
        today:
          type: number
        thisWeek:
          type: number
        thisMonth:
          type: number
        totalDeliveries:
          type: integer
        completedToday:
          type: integer
      required:
        - today
        - thisWeek
        - thisMonth
        - totalDeliveries
        - completedToday

    # Review
...